more refactoring
removing use of odespy
