+++ pyclouds/constants.py
@@ -130,67 +130,19 @@ um_constants = {
     },
 }
 
-# r, w, T, q_v, q_r, q_l, q_i
-class Var:
-    r = 0
-    w = 1
-    T = 2
-    q_v = 3
-    q_r = 4
-    q_l = 5
-    q_i = 6
-    z = 7
-    p = 8
-    q_pr = 9  # TODO: should this really be a m_precip/m_total?
+# class MySolver(odespy.RKFehlberg):
 
-    names = ['r', 'w', 'T', 'q_v', 'q_r', 'q_l', 'q_i', 'z', 'p', 'q_pr']
-    NUM = len(names)
+    # def advance(self):
+        # f, n, rtol, atol = self.f, self.n, self.rtol, self.atol
+        # u_n, t_n, t_np1 = self.u[n], self.t[n], self.t[n+1]
+        # dt = t_np1 - t_n
 
-    @staticmethod
-    def print_formatted(v, formatting='%g'):
-        print(",\t".join([("%s=" + formatting) % (Var.names[i], v[i]) for i in range(Var.NUM)]))
+        # dt_max__all = np.abs(f(u_n, t_n)/u_n)
+        # dt_max__all[REQUIRED_POSITIVE == 0] = 1.0e31
 
-    @staticmethod
-    def repr(v, formatting='%g', skip=[]):
-        units = { 'w': 'm/s', 'r': 'm', 'T': 'K', 'z': 'm', 'p': 'Pa',
-                }
-        return ", ".join([r"$%s=%g%s$" % (Var.names[i], v[i], units.get(Var.names[i], '')) for i in range(Var.NUM) if not Var.names[i] in skip])
-
-    @staticmethod
-    def make_state(**kwargs):
-        s = np.zeros((Var.NUM))
-        for k, v in list(kwargs.items()):
-            s[getattr(Var, k)] = v
-
-        return s
-
-class AttrDict(dict):
-    def __init__(self, *args, **kwargs):
-        super(AttrDict, self).__init__(*args, **kwargs)
-        self.__dict__ = self
-
-
-
-REQUIRED_POSITIVE = np.zeros((Var.NUM))
-# REQUIRED_POSITIVE[Var.r] = 1
-# REQUIRED_POSITIVE[Var.q_v] = 1
-REQUIRED_POSITIVE[Var.q_l] = 1
-# REQUIRED_POSITIVE[Var.q_r] = 1
-# REQUIRED_POSITIVE[Var.q_i] = 1
-
-class MySolver(odespy.RKFehlberg):
-
-    def advance(self):
-        f, n, rtol, atol = self.f, self.n, self.rtol, self.atol
-        u_n, t_n, t_np1 = self.u[n], self.t[n], self.t[n+1]
-        dt = t_np1 - t_n
-
-        dt_max__all = np.abs(f(u_n, t_n)/u_n)
-        dt_max__all[REQUIRED_POSITIVE == 0] = 1.0e31
-
-        self.max_step = min(np.min(dt_max__all), self.first_step)
-        print(self.max_step, dt_max__all)
-        self.first_step = self.max_step
-        import ipdb
-        with ipdb.launch_ipdb_on_exception():
-            return odespy.RKFehlberg.advance(self)
+        # self.max_step = min(np.min(dt_max__all), self.first_step)
+        # print(self.max_step, dt_max__all)
+        # self.first_step = self.max_step
+        # import ipdb
+        # with ipdb.launch_ipdb_on_exception():
+            # return odespy.RKFehlberg.advance(self)
