+++ requirements.txt
@@ -1,2 +1,1 @@
 numpy
--e git://github.com/rajeshrinet/odespy.git#egg=odespy
